Create EnemyController.cs
Added EnemyController - it does look for a target object , if target object is within a given range , it starts to look at him. If the angle between foward vector and the target is very small (which means that we're almost directly lookin at the target object) the object is moving forward. it'll stop if the distance is within a given range.